FinAgents/agent_pools/execution_agent_demo/execution_agent_demo/requirements.txt
openai-agents
alpaca-py
pandas
numpy
requests
python-dotenv

